+++ website/js/examples/geometry/clamp.js
@@ -6,9 +6,9 @@ window.onload = function() {
   var hWidth  = width  * 0.5;
   var hHeight = height * 0.5;
 
-  var circle    = TYPE6JS.Geometry.Circle.create( 0, 0, 40 );
-  var rect      = TYPE6JS.Geometry.Rectangle.create( hWidth, hHeight, hWidth, hHeight );
-  var innerRect = TYPE6JS.Geometry.Rectangle.create( hWidth, hHeight, hWidth - circle.getDiameter(), hHeight - circle.getDiameter());
+  var circle    = TYPE6.Geometry.Circle.create( 0, 0, 40 );
+  var rect      = TYPE6.Geometry.Rectangle.create( hWidth, hHeight, hWidth, hHeight );
+  var innerRect = TYPE6.Geometry.Rectangle.create( hWidth, hHeight, hWidth - circle.getDiameter(), hHeight - circle.getDiameter());
 
   //initial rendering
   drawRectangles();
@@ -20,25 +20,14 @@ window.onload = function() {
     context.clearRect(0, 0, width, height);
 
     drawRectangles();
-    drawCircle(circle, "#CC0000");
+    circle.draw( context, "#CC0000" );
 
     context.fill();
   });
 
   function drawRectangles(){
-    drawRectangle(rect, "#cccccc");
-    drawRectangle(innerRect, "#999999");
-  }
-
-  function drawRectangle(rectangle, color){
-    context.fillStyle = color;
-    context.fillRect(rectangle.getTopLeftCornerX(), rectangle.getTopLeftCornerY(), rectangle.getSizeX(), rectangle.getSizeY());
-  }
-
-  function drawCircle(circle, color){
-    context.fillStyle = color;
-    context.beginPath();
-    context.arc(circle.getPositionX(), circle.getPositionY(), circle.getRadius(), 0, TYPE6JS.Trigonometry.TWOPI, false);
+    rect.draw( context, "#cccccc" );
+    innerRect.draw( context, "#999999" );
   }
 
 };