+++ website/js/examples/random/random.js
@@ -7,26 +7,14 @@ window.onload = function() {
 
   for (var i = 0 ; i < 300 ; i += 1){
     var iMod3 = i%3;
-    var circle = TYPE6JS.Geometry.Circle.create(
-                  TYPE6JS.Random.float( width * iMod3/3, width * (iMod3+1)/3 ),
-                  TYPE6JS.Random.float( 0, height ),
-                  TYPE6JS.Random.float( 10, 40 )
+    var circle = TYPE6.Geometry.Circle.create(
+                  TYPE6.Random.float( width * iMod3/3, width * (iMod3+1)/3 ),
+                  TYPE6.Random.float( 0, height ),
+                  TYPE6.Random.float( 10, 40 )
                 );
 
-    drawCircle(circle, colors[ iMod3 ]);
+    circle.draw( context, colors[ iMod3 ] );
 
-  }
-
-  function drawCircle(circle, color){
-    context.beginPath();
-    context.fillStyle = color;
-    context.arc(
-      circle.getPositionX(),
-      circle.getPositionY(),
-      circle.getRadius(),
-      0, TYPE6JS.Trigonometry.TWOPI, false
-    );
-    context.fill();
   }
 
 };