+++ website/js/examples/trigonometry/sinewave-animation.js
@@ -5,7 +5,7 @@ window.onload = function() {
   var height  = canvas.height = window.innerHeight;
 
   var baseRadius = 50;
-  var circle     = TYPE6JS.Geometry.Circle.create( width * 0.5, 0, baseRadius );
+  var circle     = TYPE6.Geometry.Circle.create( width * 0.5, 0, baseRadius );
 
   var alpha      = 0.0;
   var baseAlpha  = 0.5;
@@ -19,24 +19,17 @@ window.onload = function() {
 
   function render() {
     //amplitude, period, shiftX, shiftY
-    alpha  = TYPE6JS.Trigonometry.sineEquation(amplitude, period, 0, baseAlpha);
-    circle.setPositionY(TYPE6JS.Trigonometry.sineEquation(amplitude * 600, period, 0, height * 0.5));
-    circle.setRadius(TYPE6JS.Trigonometry.sineEquation(amplitude * 20, period, 0, baseRadius));
+    alpha = TYPE6.Trigonometry.sineEquation(amplitude, period, 0, baseAlpha);
+    circle.setPositionY(TYPE6.Trigonometry.sineEquation(amplitude * 600, period, 0, height * 0.5));
+    circle.setRadius(TYPE6.Trigonometry.sineEquation(amplitude * 20, period, 0, baseRadius));
 
     context.clearRect(0, 0, width, height);
 
-    draw();
+    circle.draw( context, 'rgba(0, 0, 0, ' + alpha + ')' );
 
     period += speed;
 
     requestAnimationFrame( render );
-  }
-
-  function draw(){
-    context.fillStyle = "rgba(0, 0, 0, " + alpha + ")";
-    context.beginPath();
-    context.arc( circle.getPositionX(), circle.getPositionY(), circle.getRadius(), 0, TYPE6JS.Trigonometry.TWOPI, false );
-    context.fill();
   }
 
 };