+++ website/js/examples/trigonometry/spinner.js
@@ -5,21 +5,21 @@ window.onload = function() {
   var width   = canvas.width = window.innerWidth;
   var height  = canvas.height = window.innerHeight;
 
-  var mainCircle   = TYPE6JS.Geometry.Circle.create( width * 0.5, height * 0.5, 200 );
+  var mainCircle   = TYPE6.Geometry.Circle.create( width * 0.5, height * 0.5, 200 );
   var smallCircles = [];
   var angle        = 0;
   var numObjects   = 24;
   var step         = 0;
   var minAlpha     = 0.2;
-  var slice        = TYPE6JS.Trigonometry.TWOPI / numObjects;
+  var slice        = TYPE6.Trigonometry.TWOPI / numObjects;
   var radius       = 20;
-  //var circlePosition  = TYPE6JS.Vector2.create();
+  //var circlePosition  = TYPE6.Vector2.create();
 
   for(var i = 0; i < numObjects; i += 1) {
     angle = i * slice;
-    smallCircles[i] = TYPE6JS.Geometry.Circle.create(
-      TYPE6JS.Trigonometry.cosineEquation( mainCircle.getRadius(), angle, 0, mainCircle.getPositionX() ),
-      TYPE6JS.Trigonometry.sineEquation( mainCircle.getRadius(), angle, 0, mainCircle.getPositionY() ),
+    smallCircles[i] = TYPE6.Geometry.Circle.create(
+      TYPE6.Trigonometry.cosineEquation( mainCircle.getRadius(), angle, 0, mainCircle.getPositionX() ),
+      TYPE6.Trigonometry.sineEquation( mainCircle.getRadius(), angle, 0, mainCircle.getPositionY() ),
       20
     );
     var circle = smallCircles[i];
@@ -50,7 +50,7 @@ window.onload = function() {
       }else {
         circle.majAlpha();
       }
-      draw(smallCircles[i]);
+      circle.draw( context, 'rgba(0, 200, 100, ' + circle.alpha + ')');
     }
 
     step += 0.25;
@@ -59,12 +59,5 @@ window.onload = function() {
     }
     requestAnimationFrame( render );
 
-  }
-
-  function draw(circle){
-    context.fillStyle = "rgba(0, 200, 100, " + circle.alpha + ")";
-    context.beginPath();
-    context.arc( circle.getPositionX(), circle.getPositionY(), circle.getRadius(), 0, TYPE6JS.Trigonometry.TWOPI, false);
-    context.fill();
   }
 };